# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Anime Gallery Media Server",
    "description": "\nAnime Gallery Media Server API 覆盖前台图片浏览、后台媒体资料维护、批量上传任务、\n元数据导入、部署设置与系统健康检查。\n\n鉴权模型：\n\n- 浏览器后台会话使用 `agms_admin_session` HttpOnly Cookie。\n- 运维脚本建议使用 `Authorization: Bearer <AGMS_API_KEYS value>`，并只授予任务实际需要的权限。\n- 使用 Cookie 会话调用非安全方法时，还需要提交与 `agms_admin_csrf` Cookie 匹配的 `X-CSRF-Token`。\n- API Key 不依赖 Cookie，因此不需要 CSRF；可在后台独立创建、设置到期时间、调整权限或立即撤销。\n- 旧版本升级后，已有 API Key 默认迁移为全部权限，避免现有自动化任务中断。\n",
    "version": "0.1.5"
  },
  "servers": [
    {
      "url": "/",
      "description": "当前部署域名"
    },
    {
      "url": "/api",
      "description": "API 前缀"
    }
  ],
  "paths": {
...
    "/api/imports/metadata": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "预览或提交元数据导入",
        "operationId": "import_metadata_api_imports_metadata_post",
        "security": [
          {
            "OperationsApiKey": []
          },
          {
            "AdminSessionCookie": [],
            "CsrfTokenCookie": [],
            "CsrfTokenHeader": []
          }
        ],
        "parameters": [
          {
            "name": "dry_run",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Dry Run"
            },
            "description": "为 true 时只校验和预览导入变更，不正式写入。"
          },
          {
            "name": "agms_admin_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agms Admin Session"
            },
            "description": "管理员会话 Cookie。运维脚本建议改用 `Authorization: Bearer <api-key>`。"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_metadata_api_imports_metadata_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetadataImportSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "需要鉴权，或提交的令牌无效。",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Authentication required"
                }
              }
            }
          },
          "403": {
            "description": "API Key 权限不足，或 Cookie 会话发起写入请求时缺少匹配的 CSRF 请求头。",
            "content": {
              "application/json": {
                "example": {
                  "detail": "CSRF validation failed"
                }
              }
            }
          }
        },
        "description": "从 CSV、JSON、XLSX 或 XLSM 导入作品和角色。正式提交前建议先使用 `dry_run=true` 预览。\n\n所需 API Key 权限：`library:write`。",
        "x-api-key-scopes": [
          "library:write"
        ]
      }
    },
    "/api/imports/metadata/template": {
      "get": {
        "tags": [
          "imports"
        ],
        "summary": "下载元数据导入模板",
        "operationId": "download_metadata_template_api_imports_metadata_template_get",
        "security": [
          {
            "OperationsApiKey": []
          },
          {
            "AdminSessionCookie": [],
            "CsrfTokenCookie": [],
            "CsrfTokenHeader": []
          }
        ],
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "csv",
                "json",
                "xlsx",
                "xlsm"
              ],
              "type": "string",
              "default": "xlsx",
              "title": "Format"
            },
            "description": "模板格式：`csv`、`json`、`xlsx` 或 `xlsm`。"
          },
          {
            "name": "agms_admin_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agms Admin Session"
            },
            "description": "管理员会话 Cookie。运维脚本建议改用 `Authorization: Bearer <api-key>`。"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "需要鉴权，或提交的令牌无效。",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Authentication required"
                }
              }
            }
          },
          "403": {
            "description": "API Key 权限不足，或 Cookie 会话发起写入请求时缺少匹配的 CSRF 请求头。",
            "content": {
              "application/json": {
                "example": {
                  "detail": "CSRF validation failed"
                }
              }
            }
          }
        },
        "description": "下载包含全部支持字段的空白 CSV、JSON、XLSX 或 XLSM 模板。\n\n所需 API Key 权限：`library:write`。",
        "x-api-key-scopes": [
          "library:write"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_import_metadata_api_imports_metadata_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_import_metadata_api_imports_metadata_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MetadataImportRow": {
        "properties": {
          "row_number": {
            "type": "integer",
            "title": "Row Number"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "work_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Work Name"
          },
          "character_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Character Name"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "row_number",
          "action",
          "status"
        ],
        "title": "MetadataImportRow"
      },
      "MetadataImportSummary": {
        "properties": {
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run"
          },
          "total_rows": {
            "type": "integer",
            "title": "Total Rows"
          },
          "valid_rows": {
            "type": "integer",
            "title": "Valid Rows"
          },
          "error_rows": {
            "type": "integer",
            "title": "Error Rows"
          },
          "works_created": {
            "type": "integer",
            "title": "Works Created",
            "default": 0
          },
          "works_updated": {
            "type": "integer",
            "title": "Works Updated",
            "default": 0
          },
          "characters_created": {
            "type": "integer",
            "title": "Characters Created",
            "default": 0
          },
          "characters_updated": {
            "type": "integer",
            "title": "Characters Updated",
            "default": 0
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/MetadataImportRow"
            },
            "type": "array",
            "title": "Rows"
          }
        },
        "type": "object",
        "required": [
          "dry_run",
          "total_rows",
          "valid_rows",
          "error_rows"
        ],
        "title": "MetadataImportSummary"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OperationsApiKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "AGMS_API_KEY",
        "description": "后台创建的可分权运维 API Key。权限：`library:read` 媒体库读取；`uploads:manage` 图片上传；`library:write` 资料维护；`library:delete` 永久删除；`system:read` 系统只读；`settings:manage` 系统设置；`updates:read` 更新只读；`updates:run` 执行更新；`api_keys:manage` API Key 管理"
      },
      "AdminSessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "agms_admin_session",
        "description": "浏览器管理员会话 Cookie。"
      },
      "CsrfTokenCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "agms_admin_csrf",
        "description": "管理员会话同时签发的 CSRF Cookie。"
      },
      "CsrfTokenHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "使用 Cookie 会话调用非安全方法时，必须与 CSRF Cookie 一致。"
      }
    }
  },
...
}